pyproject.toml
[project]
name = "saarvis"
version = "0.1.0"
description = "Twitch-Chatbot mit OpenAI- und ElevenLabs-TTS-Integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "twitchio>=2.10.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]
[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
